Update versions to work with clap-verbosity-flag
clap-verbosity-flag depends on structopt 0.3
(https://crates.io/crates/clap-verbosity-flag), while this package used
structopt 0.2.10.

When using clap-verbosity-flag with clap-log-flag, they bring in
different versions of StructOpt, which are incompatible.

To fix this, depend on the same version of structopt as
clap-verbosity-flag, and use a recent version of clap-verbosity-flag in
dev-dependencies.